Cleaned up battle UI code. Can now be launched by a single function call, provided you pass the player and enemy (with parties)
