+++ GobalGameJam2019/Assets/Scripts/Battle/TestBattleStart.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 94c05ec99897def4e8fa0468cb877d76
+guid: c6e93f9ffa2a96f498207de2e008f4b2
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
